Move the jsonapi package to pkg/
