+++ web/remote/remote.go
@@ -2,7 +2,7 @@ package remote
 
 import (
 	"github.com/cozy/cozy-stack/pkg/remote"
-	"github.com/cozy/cozy-stack/web/jsonapi"
+	"github.com/cozy/cozy-stack/pkg/jsonapi"
 	"github.com/cozy/cozy-stack/web/middlewares"
 	"github.com/cozy/cozy-stack/web/permissions"
 	"github.com/cozy/echo"
